The Falling Platform should work now
It teleports back up. You do need to set the teleport position however. I added an empty gameobject to our green platforms which is the teleport position. You also need to add that and the falling platform to any of our green platforms as they have a script that requires it. If you have any questions feel free to ask.